+++ src/types/transform.rs
@@ -7,7 +7,7 @@ use std::ops::Mul;
 use std::f64;
 
 use {WriteOptions, WriteBuffer, WriteToString};
-use super::number::{write_num, FuzzyEq, DEFAULT_PRECISION};
+use super::number::{write_num, FuzzyEq};
 
 #[cfg(feature = "parsing")]
 use FromStream;
@@ -264,44 +264,42 @@ impl WriteBuffer for Transform {
 
 fn write_matrix_transform(ts: &Transform, opt: &WriteOptions, out: &mut Vec<u8>) {
     let rm = opt.remove_leading_zero;
-    let dp = DEFAULT_PRECISION;
 
     out.extend_from_slice(b"matrix(");
-    write_num(&ts.a, dp, rm, out);
+    write_num(&ts.a, rm, out);
     out.push(b' ');
-    write_num(&ts.b, dp, rm, out);
+    write_num(&ts.b, rm, out);
     out.push(b' ');
-    write_num(&ts.c, dp, rm, out);
+    write_num(&ts.c, rm, out);
     out.push(b' ');
-    write_num(&ts.d, dp, rm, out);
+    write_num(&ts.d, rm, out);
     out.push(b' ');
-    write_num(&ts.e, dp, rm, out);
+    write_num(&ts.e, rm, out);
     out.push(b' ');
-    write_num(&ts.f, dp, rm, out);
+    write_num(&ts.f, rm, out);
     out.push(b')');
 }
 
 fn write_simplified_transform(ts: &Transform, opt: &WriteOptions, out: &mut Vec<u8>) {
     let rm = opt.remove_leading_zero;
-    let dp = DEFAULT_PRECISION;
 
     if ts.is_translate() {
         out.extend_from_slice(b"translate(");
-        write_num(&ts.e, dp, rm, out);
+        write_num(&ts.e, rm, out);
 
         if ts.f.fuzzy_ne(&0.0) {
             out.push(b' ');
-            write_num(&ts.f, dp, rm, out);
+            write_num(&ts.f, rm, out);
         }
 
         out.push(b')');
     } else if ts.is_scale() {
         out.extend_from_slice(b"scale(");
-        write_num(&ts.a, dp, rm, out);
+        write_num(&ts.a, rm, out);
 
         if ts.a.fuzzy_ne(&ts.d) {
             out.push(b' ');
-            write_num(&ts.d, dp, rm, out);
+            write_num(&ts.d, rm, out);
         }
 
         out.push(b')');
@@ -312,7 +310,7 @@ fn write_simplified_transform(ts: &Transform, opt: &WriteOptions, out: &mut Vec<
 
         if a.fuzzy_eq(&skx) && a.fuzzy_eq(&sky) && sx.fuzzy_eq(&1.0) && sy.fuzzy_eq(&1.0) {
             out.extend_from_slice(b"rotate(");
-            write_num(&a, dp, rm, out);
+            write_num(&a, rm, out);
             out.push(b')');
         } else {
             write_matrix_transform(ts, opt, out);